Combine link parser into one file
It can always be split later if additions make it cumbersome.
